pr: alignment
